Started moving API routes to dedicated files - /v1/templates
