+++ package.json
@@ -51,12 +51,12 @@
         "@phc/pbkdf2": "1.1.14",
         "@postalsys/certs": "1.0.5",
         "@postalsys/email-ai-tools": "1.2.0",
-        "@postalsys/email-text-tools": "2.1.0",
+        "@postalsys/email-text-tools": "2.1.1",
         "@postalsys/hecks": "3.0.0-fork.3",
         "@postalsys/templates": "1.0.5",
         "ace-builds": "1.24.1",
         "base32.js": "0.1.0",
-        "bull-arena": "3.30.4",
+        "bull-arena": "4.0.0",
         "bullmq": "4.8.0",
         "compare-versions": "6.1.0",
         "dotenv": "16.3.1",
@@ -77,7 +77,7 @@
         "ioredfour": "1.2.0-ioredis-07",
         "ioredis": "5.3.2",
         "ipaddr.js": "2.1.0",
-        "joi": "17.10.0",
+        "joi": "17.10.1",
         "jquery": "3.7.1",
         "js-beautify": "1.14.9",
         "libbase64": "1.2.1",
@@ -86,7 +86,7 @@
         "license-checker": "25.0.1",
         "mailparser": "3.6.5",
         "mailsplit": "5.4.0",
-        "marked": "7.0.5",
+        "marked": "8.0.0",
         "minimist": "1.2.8",
         "msgpack5": "6.0.2",
         "murmurhash": "2.0.1",
@@ -103,7 +103,7 @@
         "smtp-server": "3.13.0",
         "socks": "2.7.1",
         "speakeasy": "2.0.0",
-        "startbootstrap-sb-admin-2": "^3.3.7",
+        "startbootstrap-sb-admin-2": "3.3.7",
         "timezones-list": "3.0.2",
         "undici": "5.23.0",
         "uuid": "9.0.0",
